requirements.txt
#MCP frameworks
fastmcp
mcp

#LLM
groq

#Environment variables
python-dotenv